helper for authentication is added and removing comments

+++ src/csrs/service/csrsService.ts
@@ -21,20 +21,11 @@ export class CsrsService {
 	async editDescription(data: any, user: any) {
 		return await this.restService.postWithoutFollowingWithConfig('api/quiz/update',
 			data,
-			{
-				headers: {
-					Authorization: `Bearer ${user.accessToken}`,
-				},
-			})
+			this.getAuthorizationHeader(user))
 	}
 
 	async editQuestion(question: any, user: any) {
-		return await this.restService.postWithoutFollowingWithConfig('api/questions/update', question,
-{
-			headers: {
-				Authorization: `Bearer ${user.accessToken}`,
-			},
-		})
+		return await this.restService.postWithoutFollowingWithConfig('api/questions/update', question, this.getAuthorizationHeader(user))
 	}
 
 	async getOrganisations() {
@@ -48,87 +39,60 @@ export class CsrsService {
 	async createQuizByProfessionID(data: any, user: any) {
 		return await this.restService.postWithoutFollowingWithConfig('/api/quiz',
 			data,
-			{
-				headers: {
-					Authorization: `Bearer ${user.accessToken}`,
-				},
-			}
+			this.getAuthorizationHeader(user)
 		)
 	}
 
 	async deleteQuizByProfession(id: number, user: any): Promise<void> {
-		await this.restService.deleteWithConfig(`/api/quiz/delete?professionId=${id}`, {
-			headers: {
-				Authorization: `Bearer ${user.accessToken}`,
-			},
-		})
+		await this.restService.deleteWithConfig(`/api/quiz/delete?professionId=${id}`, this.getAuthorizationHeader(user))
 	}
 
 	async deleteQuestionbyID(id: string, user: any): Promise<void> {
-		await this.restService.deleteWithConfig(`/api/questions/${id}/delete`, {
-			headers: {
-				Authorization: `Bearer ${user.accessToken}`,
-			},
-		})
+		await this.restService.deleteWithConfig(`/api/questions/${id}/delete`, this.getAuthorizationHeader(user))
 	}
 
 	async postQuestion(data: any, user: any) {
 		return await this.restService.postWithoutFollowingWithConfig('/api/questions/add-question',
 			data,
-			{
-				headers: {
-					Authorization: `Bearer ${user.accessToken}`,
-				},
-			}
+			this.getAuthorizationHeader(user)
 		)
 	}
 
 	async  getQuestionbyID(questionID: any, user: any) {
-		return await this.restService.getWithConfig(`/api/questions/${questionID}/preview`, {
-			headers: {
-				Authorization: `Bearer ${user.accessToken}`,
-			},
-		})
+		return await this.restService.getWithConfig(`/api/questions/${questionID}/preview`, this.getAuthorizationHeader(user))
 	}
 
 	async getResultsByProfession(id: any, user: any) {
-		return await this.restService.getWithConfig(`/api/quiz/results-by-profession?professionId=${id}`, {
-			headers: {
-				Authorization: `Bearer ${user.accessToken}`,
-			},
-		})
+		return await this.restService.getWithConfig(`/api/quiz/results-by-profession?professionId=${id}`, this.getAuthorizationHeader(user))
 	}
 
-	async publishSkills(profession: any) {
-		return await this.restService.put(`/api/quiz/publish`, {profession: profession})
+	async publishSkills(profession: any, user: any) {
+		return await this.restService.put(`/api/quiz/publish`, {profession: profession}, )
 	}
 
 	async getQuizByProfession(id: any, user: any) {
-		return await this.restService.getWithConfig(`/api/quiz/${id}`, {
-			headers: {
-				Authorization: `Bearer ${user.accessToken}`,
-			},
-		})
+		return await this.restService.getWithConfig(`/api/quiz/${id}`, this.getAuthorizationHeader(user))
 	}
 
 	async getAllQuizResults(user: any) {
-		return await this.restService.getWithConfig(`/api/quiz/all-results`, {
-			headers: {
-				Authorization: `Bearer ${user.accessToken}`,
-			},
-		})
+		return await this.restService.getWithConfig(`/api/quiz/all-results`, this.getAuthorizationHeader(user))
 	}
 
 	async getQuizesByOrg(orgID: any, user: any) {
-		return await this.restService.getWithConfig(`api/quiz/results-for-your-org?organisationId=${orgID}`, {
+		return await this.restService.getWithConfig(`api/quiz/results-for-your-org?organisationId=${orgID}`, this.getAuthorizationHeader(user))
+	}
+
+	private getAuthorizationHeader(user: any) {
+		return {
 			headers: {
 				Authorization: `Bearer ${user.accessToken}`,
 			},
-		})
+		}
 	}
 
 	async getAreasOfWork() {
 		let areasOfWork = await this.restService.get('/professions/flat')
+
 		return areasOfWork
 	}
